refactor: remove internal mod exports
