+++ src/databases/sqlite.rs
@@ -8,8 +8,8 @@ use r2d2_sqlite::SqliteConnectionManager;
 use crate::databases::database;
 use crate::databases::database::{Database, Error};
 use crate::protocol::clock::DurationSinceUnixEpoch;
+use crate::protocol::common::InfoHash;
 use crate::tracker::key::AuthKey;
-use crate::InfoHash;
 
 pub struct SqliteDatabase {
     pool: Pool<SqliteConnectionManager>,
