+++ src/lib.rs
@@ -1,11 +1,3 @@
-pub use api::server::*;
-pub use http::server::*;
-pub use protocol::common::*;
-pub use udp::server::*;
-
-pub use self::config::*;
-pub use self::tracker::*;
-
 pub mod api;
 pub mod config;
 pub mod databases;
