+++ src/setup.rs
@@ -3,9 +3,9 @@ use std::sync::Arc;
 use log::warn;
 use tokio::task::JoinHandle;
 
+use crate::config::Configuration;
 use crate::jobs::{http_tracker, torrent_cleanup, tracker_api, udp_tracker};
 use crate::tracker::TorrentTracker;
-use crate::Configuration;
 
 pub async fn setup(config: &Configuration, tracker: Arc<TorrentTracker>) -> Vec<JoinHandle<()>> {
     let mut jobs: Vec<JoinHandle<()>> = Vec::new();
